Revert "refactor GetDeviceInfoFromPath method from devices pkg to vmutils"
This reverts commit 7ece36daed0d67f5c0c155b77278eb1e6a78ecd2.
