+++ internal/oci/uvm.go
@@ -16,6 +16,7 @@ import (
 	"github.com/sirupsen/logrus"
 
 	iannotations "github.com/Microsoft/hcsshim/internal/annotations"
+	"github.com/Microsoft/hcsshim/internal/devices"
 	"github.com/Microsoft/hcsshim/internal/log"
 	"github.com/Microsoft/hcsshim/internal/uvm"
 )
@@ -285,7 +286,7 @@ func parseDevices(ctx context.Context, specWindows *specs.Windows) []uvm.VPCIDev
 	}
 	extraDevices := []uvm.VPCIDeviceID{}
 	for _, d := range specWindows.Devices {
-		pciID, index := vmutils.GetDeviceInfoFromPath(d.ID)
+		pciID, index := devices.GetDeviceInfoFromPath(d.ID)
 		if vmutils.IsValidDeviceType(d.IDType) {
 			key := uvm.NewVPCIDeviceID(pciID, index)
 			extraDevices = append(extraDevices, key)
