+++ internal/vm/vmutils/utils.go
@@ -8,7 +8,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"strconv"
 
 	"github.com/Microsoft/hcsshim/internal/log"
 )
@@ -27,19 +26,6 @@ func IsValidDeviceType(deviceType string) bool {
 	return (deviceType == VPCIDeviceIDType) ||
 		(deviceType == VPCIDeviceIDTypeLegacy) ||
 		(deviceType == GpuDeviceIDType)
-}
-
-// GetDeviceInfoFromPath takes a device path and parses it into the PCI ID and
-// virtual function index if one is specified.
-func GetDeviceInfoFromPath(rawDevicePath string) (string, uint16) {
-	indexString := filepath.Base(rawDevicePath)
-	index, err := strconv.ParseUint(indexString, 10, 16)
-	if err == nil {
-		// we have a vf index
-		return filepath.Dir(rawDevicePath), uint16(index)
-	}
-	// otherwise, just use default index and full device ID given
-	return rawDevicePath, 0
 }
 
 // ParseUVMReferenceInfo reads the UVM reference info file, and base64 encodes the content if it exists.
